// front/src/app/swagger/model/productResponseDto.ts
/**
 * ExamenTec API
 *
 * Contact: support@examentec.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductResponseDto { 
    id?: string;
    name?: string | null;
    description?: string | null;
    price?: number;
    stock?: number;
    categoryId?: string;
    categoryName?: string | null;
    storeName?: string | null;
    createdDate?: string;
}