// front/src/app/swagger/model/productResponseDtoPagedResult.ts
/**
 * ExamenTec API
 *
 * Contact: support@examentec.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductResponseDto } from './productResponseDto';


export interface ProductResponseDtoPagedResult { 
    items?: Array<ProductResponseDto> | null;
    page?: number;
    pageSize?: number;
    totalCount?: number;
    readonly totalPages?: number;
    readonly hasPreviousPage?: boolean;
    readonly hasNextPage?: boolean;
}